refactor: split test_tools_github.py — extract report_bug tests into test_tools_report_bug.py
Moves all report_bug handler tests and supporting private-helper unit
tests (_parse_fingerprint, _extract_block, _parse_prepare_result,
_parse_enrich_result) into a new focused file (~680 lines). The original
file retains only fetch_github_issue and get_issue_title tests (~200 lines).
Updates tests/server/CLAUDE.md with the new file entry.
